Map overloaded in DbQuestionMapper.cs, deleted GroupQuestionMapper
