+++ src/SurveyService.Business/Commands/Question/CreateQuestionCommand.cs
@@ -17,14 +17,14 @@ namespace LT.DigitalOffice.SurveyService.Business.Commands.Question;
 public class CreateQuestionCommand : ICreateQuestionCommand
 {
   private readonly IHttpContextAccessor _httpContextAccessor;
-  private readonly IDbSingleQuestionMapper _mapper;
+  private readonly IDbQuestionMapper _mapper;
   private readonly IQuestionRepository _questionRepository;
   private readonly ICreateSingleQuestionRequestValidator _validator;
   private readonly IResponseCreator _responseCreator;
 
   public CreateQuestionCommand(
     IHttpContextAccessor httpContextAccessor, 
-    IDbSingleQuestionMapper mapper, 
+    IDbQuestionMapper mapper, 
     IQuestionRepository questionRepository,
     ICreateSingleQuestionRequestValidator validator,
     IResponseCreator responseCreator)
