+++ src/SurveyService.Mappers/Db/DbQuestionMapper.cs
@@ -8,15 +8,20 @@ using System.Linq;
 
 namespace LT.DigitalOffice.SurveyService.Mappers.Db;
 
-public class DbSingleQuestionMapper : IDbSingleQuestionMapper
+public class DbQuestionMapper : IDbQuestionMapper
 {
   private readonly IHttpContextAccessor _httpContextAccessor;
   private readonly IDbOptionMapper _dbOptionMapper;
+  private readonly IDbQuestionMapper _dbQuestionMapper;
 
-  public DbSingleQuestionMapper(IHttpContextAccessor httpContextAccessor, IDbOptionMapper dbOptionMapper)
+  public DbQuestionMapper(
+    IHttpContextAccessor httpContextAccessor, 
+    IDbOptionMapper dbOptionMapper,
+    IDbQuestionMapper dbSingleQuestionMapper)
   {
     _httpContextAccessor = httpContextAccessor;
     _dbOptionMapper = dbOptionMapper;
+    _dbQuestionMapper = dbSingleQuestionMapper;
   }
 
   public DbQuestion Map(CreateSingleQuestionRequest request)
@@ -25,7 +30,7 @@ public class DbSingleQuestionMapper : IDbSingleQuestionMapper
 
     return request is null
     ? null
-    : new DbQuestion()
+    : new DbQuestion
     {
       Id = questionId,
       GroupId = request.GroupId,
@@ -43,5 +48,26 @@ public class DbSingleQuestionMapper : IDbSingleQuestionMapper
       CreatedAtUtc = DateTime.UtcNow,
       Options = request.Options?.Select(option => _dbOptionMapper.Map(option, questionId)).ToList()
     };
+  }
+  
+  public DbQuestion Map(CreateGroupQuestionRequest request, Guid groupId, DateTime? groupDeadline, bool groupHasRealTimeResult)
+  {
+    return request is null
+      ? null
+      : _dbQuestionMapper.Map(
+        new CreateSingleQuestionRequest 
+        {
+          GroupId = groupId,
+          Content = request.Content,
+          Deadline = groupDeadline,
+          HasRealTimeResult = groupHasRealTimeResult,
+          IsAnonymous = request.IsAnonymous,
+          IsRevoteAvailable = request.IsRevoteAvailable,
+          IsObligatory = request.IsObligatory,
+          IsPrivate = request.IsPrivate,
+          HasMultipleChoice = request.HasMultipleChoice,
+          HasCustomOptions = request.HasCustomOptions,
+          Options = request.Options
+        });
   }
 }
