--- src/SurveyService.Mappers/Db/Interfaces/IDbGroupQuestionMapper.cs
@@ -1,12 +1,0 @@
-using LT.DigitalOffice.Kernel.Attributes;
-using LT.DigitalOffice.SurveyService.Models.Db;
-using LT.DigitalOffice.SurveyService.Models.Dto.Requests.Question;
-using System;
-
-namespace LT.DigitalOffice.SurveyService.Mappers.Db.Interfaces;
-
-[AutoInject]
-public interface IDbGroupQuestionMapper
-{
-  DbQuestion Map(CreateGroupQuestionRequest request, Guid groupId, DateTime? groupDeadline, bool groupHasRealTimeResult);
-}
